SEQNG-423 Split the event model #time 30m
